+++ src/ptex/PtexHalf.cpp
@@ -73,7 +73,7 @@ static bool PtexHalfInit()
 	    int s = ((f>>16) & 0x8000);
 	    int m = f & 0x7fe000;
 	    // add bit 12 to round
-	    PtexHalf::f2hTable[i] = (s|((e|m)>>13))+((f>>12)&1);
+	    PtexHalf::f2hTable[i] = (uint16_t)((s|((e|m)>>13))+((f>>12)&1));
 	}
     }
 
@@ -93,13 +93,13 @@ uint16_t PtexHalf::fromFloat_except(uint32_t i)
 	// denormalized
 	union { uint32_t i; float f; } u;
 	u.i = i;
-	return s | int(fabs(u.f)*1.6777216e7 + .5);
+	return (uint16_t)(s|int(fabs(u.f)*1.6777216e7 + .5));
     }
 
     if (e == 0x23c00)
 	// inf/nan, preserve msb bits of m for nan code
-	return s|0x7c00|((i&0x7fffff)>>13);
+	return (uint16_t)(s|0x7c00|((i&0x7fffff)>>13));
     else
 	// overflow - convert to inf
-	return s|0x7c00;
+	return (uint16_t)(s|0x7c00);
 }
